Changed: Move permission module from agents to core crate
Relocate the permission evaluation system from llm-coding-tools-agents
to llm-coding-tools-core where it more appropriately belongs as a
framework-agnostic primitive.

Changes:
- Moved permission.rs from agents crate to core crate as permissions.rs
- Moved PermissionAction and PermissionRule enums to appropriate crates
- Core permissions module contains: PermissionAction, Rule, Ruleset, wildcard matching
- Serialization format (PermissionRule) remains in agents crate
- Updated all imports across serdesai crate and tests
- Added llm-coding-tools-core dependency to agents crate

Benefits:
- Better separation of concerns: core evaluation logic vs serialization
- Permission system is now framework-agnostic and reusable
- Agents crate focuses on config loading and agent management
- Clearer API boundaries between crates
