+++ src/llm-coding-tools-agents/src/lib.rs
@@ -58,14 +58,15 @@
 mod catalog;
 mod config;
 mod error;
+mod extensions;
 mod loader;
 mod parser;
-mod permission;
 
 pub use catalog::AgentCatalog;
-pub use config::{AgentConfig, AgentMode, PermissionAction, PermissionRule};
+pub use config::{AgentConfig, AgentMode, PermissionRule};
 pub use error::AgentLoadError;
 pub use error::AgentLoadResult;
+pub use extensions::RulesetExt;
+pub use llm_coding_tools_core::permissions::{PermissionAction, Rule, Ruleset};
 pub use loader::AgentLoader;
 pub use parser::AgentParseError;
-pub use permission::{Rule, Ruleset};
