+++ src/llm-coding-tools-core/src/lib.rs
@@ -13,6 +13,7 @@ pub mod error;
 pub mod fs;
 pub mod output;
 pub mod path;
+pub mod permissions;
 pub mod system_prompt;
 pub mod tool_names;
 pub mod tools;
