+++ src/llm-coding-tools-serdesai/src/task/mod.rs
@@ -5,7 +5,8 @@
 use crate::convert::to_serdes_result;
 use crate::registry::{AgentRegistry, RegistryAgent};
 use async_trait::async_trait;
-use llm_coding_tools_agents::{AgentMode, PermissionAction, Ruleset};
+use llm_coding_tools_agents::AgentMode;
+use llm_coding_tools_core::permissions::{PermissionAction, Ruleset};
 use llm_coding_tools_core::context::ToolContext;
 use llm_coding_tools_core::tool_names;
 use llm_coding_tools_core::tools::TaskInput;
